requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0

openai==1.12.0
httpx==0.27.0

python-dotenv==1.0.1

pytest==8.1.1
pytest-asyncio==0.23.5
